+++ src/main/java/com/deiksoftdev/automatagame/config/resourceserving/ResourceServingProperties.java
