+++ src/main/java/com/deiksoftdev/automatagame/config/resourceserving/StaticResourceConfiguration.java
