+++ src/main/java/com/deiksoftdev/automatagame/config/security/CustomAuthenticationFilter.java
