+++ src/main/java/com/deiksoftdev/automatagame/config/security/CustomAuthenticationToken.java
