+++ src/main/java/com/deiksoftdev/automatagame/dto/UserRegisterDTO.java
