+++ src/main/java/com/deiksoftdev/automatagame/exception/UserByIdNotRegisteredException.java
