+++ src/main/java/com/deiksoftdev/automatagame/exception/UserEmailAlreadyExistsException.java
