+++ src/main/java/com/deiksoftdev/automatagame/exception/UserNameAlreadyExistsException.java
