+++ src/main/java/com/deiksoftdev/automatagame/model/UserRepository.java
@@ -1,0 +1,10 @@
+package com.deiksoftdev.automatagame.model;
+
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface UserRepository extends MongoRepository<User, Long> {
+
+    User findByName(String name);
+
+    User findByEmail(String email);
+}
