--- src/main/java/com/deiksoftdev/automatagame/user/UserRepository.java
@@ -1,12 +1,0 @@
-package com.deiksoftdev.automatagame.user;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-
-    User findByName(String name);
-
-    User findByEmail(String email);
-}
