+++ src/main/java/com/deiksoftdev/automatagame/web/controller/LandingPageController.java
@@ -1,4 +1,4 @@
-package com.deiksoftdev.automatagame.landing;
+package com.deiksoftdev.automatagame.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
