+++ src/main/java/com/deiksoftdev/automatagame/web/controller/UserListController.java
@@ -1,0 +1,24 @@
+package com.deiksoftdev.automatagame.web;
+
+import com.deiksoftdev.automatagame.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class UserListController {
+
+    private final UserService userService;
+
+    @Autowired
+    public UserListController(UserService userService) {
+        this.userService = userService;
+    }
+
+    @GetMapping("/users")
+    public String showUsersList(Model model) {
+        model.addAttribute("users", userService.findAll());
+        return "user-list";
+    }
+}
