+++ src/main/java/com/deiksoftdev/automatagame/web/controller/UserLoginController.java
@@ -1,8 +1,7 @@
-package com.deiksoftdev.automatagame.login;
+package com.deiksoftdev.automatagame.web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
-
 
 @Controller
 public class UserLoginController {
